refactored app tree
