+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.F32Test"
         tools:targetApi="31">
         <activity
-            android:name=".ui.main.MainActivity"
+            android:name="app.seals.f32test.main.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.F32Test"
