+++ app/src/main/java/app/seals/f32test/di/uiDi.kt
@@ -1,6 +1,6 @@
 package app.seals.f32test.di
 
-import app.seals.f32test.ui.main.vm.MainActivityViewModel
+import app.seals.f32test.main.vm.MainActivityViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
