+++ app/src/main/java/app/seals/f32test/main/vm/MainActivityViewModel.kt
@@ -1,8 +1,6 @@
-package app.seals.f32test.ui.main.vm
+package app.seals.f32test.main.vm
 
-import android.util.Log
 import androidx.compose.foundation.lazy.LazyListState
-import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.lifecycle.viewModelScope
 import app.seals.f32test.ui.sampledata.DataPump
 import kotlinx.coroutines.launch
