+++ app/src/main/java/app/seals/f32test/ui/navigation/NavigationGraph.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import app.seals.f32test.ui.main.vm.MainActivityViewModel
+import app.seals.f32test.main.vm.MainActivityViewModel
 import app.seals.f32test.ui.sampledata.DataPump
 import app.seals.f32test.ui.screens.cart.CartScreen
 import app.seals.f32test.ui.screens.details.ProductDetails
